post, account api 에러처리 및 피드백 내용 반영 완료

+++ src/routers/post.js
@@ -1,32 +1,40 @@
 // Import
 const router = require("express").Router()
 const conn = require('../../database/connect/maria')
+const exception = require('../modules/exception')
 
 //Apis
 //게시글 목록(게시판)
-router.get("/", (res) => {
+router.get("/", (req, res) => {
     const postBoardResult = {
         "success": false,
-        "message": ""
+        "message": "",
+        "data": null
     }
     try{
+        if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
+        const idx = req.session.user.idx;
+
+        //db에 값 입력하기
         conn.query('SELECT * FROM post ORDER BY idx DESC', (err, results) => {
-            if (err) res.send(postBoardResult)
+            if (err) return res.send(postBoardResult)
             if (results.length > 0) {
-                postBoardResult.success = true;
-                res.send(postBoardResult);
+                postBoardResult.success = true
+                postBoardResult.data = results
+                res.send(postBoardResult)
             }
             else {
-                postBoardResult.success = true;
-                postBoardResult.message = "게시글 목록이 비어있습니다.";
+                postBoardResult.success = true
+                postBoardResult.message = "게시글 목록이 비어있습니다."
+                res.send(postBoardResult)
             }
         })
     }
     catch (e) {
-        postBoardResult.message = e.message;
-        res.status(400).send(postBoardResult);
+        postBoardResult.message = e.message
+        res.status(400).send(postBoardResult)
     }
-});
+})
 
 //게시글 업로드
 router.post("/", (req, res) => {
@@ -36,56 +44,52 @@ router.post("/", (req, res) => {
         "message": ""
     }
     try {
-        if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
-        const idx = req.session.user.idx;
+        if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.")
+        const idx = req.session.user.idx
+        exception.titleCheck(title)
+        exception.contentCheck(content)
 
         //db에 값 입력하기
         conn.query('INSERT INTO post (account_idx, title, content) VALUES (?, ?, ?)', [idx, title, content], (err) => {
-            if (err) {
-                throw new Error("데이터베이스가 이상해요")
-            }
-            else {
-                console.log("성공");
-                uploadPostResult.success = true;
-                uploadPostResult.message = "게시글이 업로드되었습니다.";
-                res.send(uploadPostResult)
-            }
-        });
+            if (err) return res.send(uploadPostResult)
+            uploadPostResult.success = true
+            res.send(uploadPostResult)
+        })
     }
     catch (e) {
-        uploadPostResult.message = e.message;
-        res.status(400).send(uploadPostResult);
+        uploadPostResult.message = e.message
+        res.status(400).send(uploadPostResult)
     }
 })
 
 //게시글 수정
 router.put("/:postidx", (req, res) => {
-    const postIdx = req.params.postidx;
-    const {title, content} = req.body;
+    const postIdx = req.params.postidx
+    const {title, content} = req.body
     const editPostResult = {
         "success": false,
         "message": ""
     }
     try {
         if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
-        const idx = req.session.user.idx;
+        const idx = req.session.user.idx
 
         //db에 값 입력하기
-        conn.query('UPDATE post SET title=?, content=? WHERE idx=?', [title, content, postIdx], (err) => {
-            if (err) {
-                throw new Error("데이터베이스가 이상해요")
+        conn.query('UPDATE post SET title=?, content=? WHERE idx=? AND account_idx=?', [title, content, postIdx, idx], (err, results) => {
+            if (err) return res.send(editPostResult)
+            if (results.affectedRows === 0) {
+                // 수정된 행이 없는 경우 처리
+                editPostResult.message = "해당하는 게시글을 찾지 못했습니다."
+                return res.send(editPostResult)
             }
-            else {
-                console.log("성공");
-                editPostResult.success = true;
-                editPostResult.message = "게시글 수정이 완료되었습니다.";
-                res.send(editPostResult)
-            }
-        });
+            editPostResult.success = true
+            editPostResult.message = "게시글 수정이 완료되었습니다."
+            res.send(editPostResult)
+        })
     }
     catch (e) {
-        editPostResult.message = e.message;
-        res.status(400).send(editPostResult);
+        editPostResult.message = e.message
+        res.status(400).send(editPostResult)
     }
 })
 
@@ -97,20 +101,21 @@ router.delete("/:postidx", (req, res) => {
         "message": ""
     }
     try {
-        if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.");
-        const idx = req.session.user.idx;
-        console.log(idx, postIdx)
+        if (!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.")
+        const idx = req.session.user.idx
+
         //db에 값 입력하기
-        conn.query('DELETE FROM post WHERE idx=?', [postIdx], (err) => {
-            if (err) {
-                throw new Error("데이터베이스가 이상해요")
-            } 
-            else {
-                deletePostResult.success = true;
-                deletePostResult.message = "게시글이 삭제되었습니다.";
-                res.send(deletePostResult);
+        conn.query('DELETE FROM post WHERE idx=? AND account_idx=?', [postIdx, idx], (err, results) => {
+            if (err) return res.send(deletePostResult)
+            if (results.affectedRows === 0) {
+                // 수정된 행이 없는 경우 처리
+                deletePostResult.message = "해당하는 게시글을 찾지 못했습니다."
+                return res.send(deletePostResult)
             }
-        });
+            deletePostResult.success = true
+            deletePostResult.message = "게시글이 삭제되었습니다."
+            res.send(deletePostResult)
+        })
     }
     catch (e) {
         deletePostResult.message = e.message;
@@ -118,41 +123,31 @@ router.delete("/:postidx", (req, res) => {
     }
 })
 
+//게시글 보기
 router.get("/:postidx", (req, res) => {
     const postIdx = req.params.postidx;
     const viewPostResult = {
         "success": false,
         "message": "",
-        "accountIdx": "",
-        "title": "",
-        "content": ""
+        "data": null
     }
     try {
         if(!req.session.user) throw new Error("세션에 사용자 정보가 없습니다.")
 
         conn.query('SELECT account_idx, title, content FROM post WHERE idx=?', [postIdx], (err,results) => {
-            if (err) {
-                throw new Error("데이터베이스가 이상해요")
+            if (err) return res.send(viewPostResult)
+            if (results.length === 0) {
+                viewPostResult.message = "게시글을 찾을 수 없습니다."
+                return res.send(viewPostResult)
             }
-            else {
-                if (results.length > 0) {
-                    viewPostResult.success = true;
-                    viewPostResult.message = "게시글 불러오기 성공";
-                    viewPostResult.accountIdx = results[0].account_idx;
-                    viewPostResult.title = results[0].title;
-                    viewPostResult.content = results[0].content;
-                    res.send(viewPostResult);
-                }
-                else {
-                    viewPostResult.message = "게시글을 찾을 수 없습니다.";
-                    res.status(404).send(viewPostResult);
-                }
-            }
+            viewPostResult.success = true;
+            viewPostResult.data = results
+            res.send(viewPostResult)
         })
     }
     catch (e) {
-        viewPostResult.message = e.message;
-        res.status(400).send(viewPostResult);
+        viewPostResult.message = e.message
+        res.status(400).send(viewPostResult)
     }
 })
